+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+### Changed
 
+Added TopCount support in QueryByAttribute, and throw exception if both TopCount and PageInfo are set: DynamicsValue/fake-xrm-easy#16
 
 ## [2.0.1-rc1] - Initial release