+++ CHANGELOG.md
@@ -3,5 +3,7 @@
 ### Changed
 
 Added TopCount support in QueryByAttribute, and throw exception if both TopCount and PageInfo are set: DynamicsValue/fake-xrm-easy#16
+Removed .netcoreapp3.1 target framework in versions 2.x, it'll be supported from versions 3.x onwards.
+Bump Microsoft.CrmSdk.CoreAssemblies to version 9.0.2.27 to support plugin telemetry - DynamicsValue/fake-xrm-easy#24
 
 ## [2.0.1-rc1] - Initial release